Renamed and improved the workspace's integrated playground
